A couple of things were in there as ints when they should have been doubles. That's been corrected.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Subsystems/Control/Control.java
@@ -159,8 +159,8 @@ public class Control extends Subsystem {
         1) https://github.com/acmerobotics/road-runner/blob/master/doc/notebook/road-runner-lite.ipynb)
      */
     public void moveLinearSlidePID(LinearSlidePosition newPosition) {
-        int lastError = 0;
-        int integral = 0;
+        double lastError = 0;
+        double integral = 0;
         double error;
         double pFix;
         double iFix;
